brand administration ready. server side

+++ server/controllers/brandController.js
@@ -5,8 +5,8 @@ const { validationResult } = require('express-validator');
 class BrandController {
     async create(req, res, next){
         try{
-            const {brand} = req.body;
-            const data = await brandService.create(brand);
+            const {brands} = req.body;
+            const data = await brandService.create(brands);
             return res.json(data);
         }catch(e){
             next(ApiError.forbidden(e.message));
@@ -14,7 +14,8 @@ class BrandController {
     }
     async getAll(req, res){
         try{
-            const brands = await brandService.getAll();
+            let { sortBy, sortDirection } = req.query;
+            const brands = await brandService.getAll(sortBy, sortDirection);
             return res.json(brands)
         }catch(e){
             next(ApiError.forbidden(e.message));
@@ -36,6 +37,10 @@ class BrandController {
     }
     async delete(req, res, next){
         try{
+            const errors = validationResult(req);
+            if(!errors.isEmpty()){
+                return next(ApiError.badRequest('validation error: ', errors.array()));
+            }
             const { id } = req.body;
             const data = await brandService.delete(id);
             return res.json(data);
